OTWO-4049 Address destroy reverification method --WIP

+++ app/models/reverification_tracker.rb
@@ -55,8 +55,9 @@ class ReverificationTracker < ActiveRecord::Base
       end
     end
 
-    def remove_reverification_trackers_for_verifed_accounts
+    def remove_reverification_trackers_for_verified_accounts
       find_each do |rev_tracker|
+        next unless rev_tracker.account.try(:access)
         rev_tracker.destroy if rev_tracker.account.access.mobile_or_oauth_verified?
       end
     end
